wip - some refactor, currency in its own file and renaming settingspickercell delegates
